Feature: Now checks to see if date "key" exists and allows you to either cancel or rewrite.

+++ main.js
@@ -8,6 +8,7 @@ $(document).ready(function() { // ready
     //  ///////  save button event ///////////
     $(save).click(function() { // save inputs to localStorage
         let date = $('#date').val();
+
         // data object of inputs
         let dayObject = {
             consumed: $('#consumed').val(),
@@ -16,12 +17,22 @@ $(document).ready(function() { // ready
         }
 
         // clear input boxes
-        // FIXME - clear input boxes
-
+        $('#date').val('');
+        $('#consumed').val('');
+        $('#burned').val('');
+        $('#hrv').val('');
 
         //add to local storage 
         let dayObjStr = JSON.stringify(dayObject);
-        // TODO - check to see if date "key" exists in localStorage first
+        if (localStorage.getItem(date)) { // check to see if date "key" exists
+            let rewrite = confirm("Date already entered. Do you want to replace?");
+            if (rewrite == false) {
+                return false;
+            } else { // rewrite data if confirmed...
+                localStorage.setItem(date, dayObjStr);
+                console.log("not entered");
+            }
+        }
         localStorage.setItem(date, dayObjStr);
 
         // Add to local storage values to table /////
